backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.17
itsdangerous==2.2.0
python-dotenv==1.0.1
httpx==0.28.1
google-generativeai==0.8.3
jinja2==3.1.4
